backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
requests
python-dotenv
google-generativeai